// case_5_on_trace/sentio/src/schema/schema.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type { String, Int, BigInt, Float, ID, Bytes, Timestamp, Boolean, Int8 } from '@sentio/sdk/store'
import { Entity, Required, One, Many, Column, ListColumn, AbstractEntity, getStore, UpdateValues } from '@sentio/sdk/store'
import { BigDecimal } from '@sentio/bigdecimal'
import { DatabaseSchema } from '@sentio/sdk'







interface SwapConstructorInput {
  id: ID;
  blockNumber: BigInt;
  transactionHash: String;
  from: String;
  to: String;
  amountIn: BigInt;
  amountOutMin: BigInt;
  deadline: BigInt;
  path: String;
  pathLength: Int;
  indexedAt: BigInt;
}
@Entity("Swap")
export class Swap extends AbstractEntity  {

	@Required
	@Column("ID")
	id: ID

	@Required
	@Column("BigInt")
	blockNumber: BigInt

	@Required
	@Column("String")
	transactionHash: String

	@Required
	@Column("String")
	from: String

	@Required
	@Column("String")
	to: String

	@Required
	@Column("BigInt")
	amountIn: BigInt

	@Required
	@Column("BigInt")
	amountOutMin: BigInt

	@Required
	@Column("BigInt")
	deadline: BigInt

	@Required
	@Column("String")
	path: String

	@Required
	@Column("Int")
	pathLength: Int

	@Required
	@Column("BigInt")
	indexedAt: BigInt
  constructor(data: SwapConstructorInput) {super()}
  
  
  static update(values: UpdateValues<SwapConstructorInput>): Promise<void> {
    return getStore().update(Swap, values)
  }
}


const source = `type Swap @entity {
  id: ID!
  blockNumber: BigInt!
  transactionHash: String!
  from: String!
  to: String!
  amountIn: BigInt!
  amountOutMin: BigInt!
  deadline: BigInt!
  path: String!
  pathLength: Int!
  indexedAt: BigInt!
}

`
DatabaseSchema.register({
  source,
  entities: {
    "Swap": Swap
  }
})